add move method

+++ src/PositionGenerator.go
@@ -6,8 +6,8 @@ import (
 )
 
 type PositionGenerator struct {
-	maxX int
-	maxY int
+	maxX      int
+	maxY      int
 	generator rand.Rand
 }
 
@@ -15,11 +15,11 @@ func NewPositionGenerator(maxX int, maxY int) PositionGenerator {
 	var seed = time.Now().UnixNano()
 	var source = rand.NewSource(seed)
 	var generator = rand.New(source)
-	return PositionGenerator { maxX, maxY, *generator }
+	return PositionGenerator{maxX, maxY, *generator}
 }
 
 func (generator *PositionGenerator) Make() Position {
 	var x = generator.generator.Intn(generator.maxX)
 	var y = generator.generator.Intn(generator.maxY)
-	return NewPosition(uint(x), uint(y))
+	return NewPosition(x, y)
 }
